requirements (1).txt
python-telegram-bot==20.7
pymongo==4.6.0
flask==2.3.3
